Use InternalRowTransferDescriptor for SparkToVectorEnginePlan InternalRow branch
